💄 Design : 실시간 알림 위치 고정 #104
